docs(ComponentRoute): Extract prop renderer
